+++ src/SeablastModelInterface.php
@@ -1,8 +1,11 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Seablast\Seablast;
 
 use Seablast\Seablast\SeablastConfiguration;
+use Seablast\Seablast\Superglobals;
 use stdClass;
 
 /**
@@ -12,6 +15,6 @@ use stdClass;
  */
 interface SeablastModelInterface
 {
-    public function __construct(SeablastConfiguration $configuration);
-    public function getParameters(): stdClass;
+    public function __construct(SeablastConfiguration $configuration, Superglobals $superglobals);
+    public function knowledge(): stdClass;
 }